refactor: Move IAsk types to @chainlit/components
